Fix bug with method getEngineById
