+++ week5/readme
@@ -4,4 +4,4 @@ Software requirements:
 
 How To:
 Init DB by creating propriate database and needed tables.
-All commands and instruction can be found at cars_catalog.
+All commands and instruction can be found at "cars_catalog".
